simple scrobbler support on winddos ( needs installed last.fm client )
update sqlite from 3.5.4 to 3.6.1

update fmodex to 4.18.1

+++ trunk/wxMusik/src/MusikApp.cpp
@@ -210,7 +210,7 @@ bool MusikApp::OnInit()
 					sData += wxT("\n");
 				}
 		#if wxUSE_UNICODE
-				pConn->Poke(wxT("PlayFiles"),sData.GetWriteBuf(sData.Length()),sData.Length()*sizeof(wxChar),wxIPC_UNICODETEXT);
+				pConn->Poke(wxT("PlayFiles"),sData.GetWriteBuf(sData.Length()),(int)sData.Length()*sizeof(wxChar),wxIPC_UNICODETEXT);
 		#else
 				pConn->Poke(wxT("PlayFiles"),sData.GetWriteBuf(sData.Length()),sData.Length(),wxIPC_TEXT);
 		#endif
